refactor(storage): consolidate migration contracts
